Remove special-case for rectangles
It's not always correct (e.g. imagine a triangle, make a copy and
rotate it 90º around the center of its bounding box. This would be
marked as a rectangle even though it's not), and writing accurate
logic slows down the render too much when you get a lot of primitives.

Rely on the texture cache instead.